+++ code/src/java/plugin/lsttokens/UmultLst.java
@@ -81,7 +81,7 @@ public class UmultLst extends AbstractIntToken<CDOMObject> implements CDOMPrimar
 		}
 		if (mult != null)
 		{
-			if (mult.intValue() <= 0)
+			if (mult <= 0)
 			{
 				context.addWriteMessage(getTokenName() + " must be an integer > 0");
 				return null;
