+++ code/src/java/plugin/primitive/spell/SpellTypeToken.java
@@ -61,7 +61,7 @@ public class SpellTypeToken extends AbstractRestrictedSpellPrimitive
 			{
 				for (Integer level : levelInfo.getListFor(spellList))
 				{
-					if (allow(pc, level.intValue(), source, spell, null))
+					if (allow(pc, level, source, spell, null))
 					{
 						return true;
 					}
